LPS-84130 Move to com.liferay.portal.kernel.servlet
